+++ code/packages/rust/ruby-parser/src/_grammar.rs
@@ -26,6 +26,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"while_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"until_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"assignment"#.to_string() },
+                GrammarElement::RuleReference { name: r#"method_with_block"#.to_string() },
                 GrammarElement::RuleReference { name: r#"method_call"#.to_string() },
                 GrammarElement::RuleReference { name: r#"expression_stmt"#.to_string() },
             ] },
@@ -76,6 +77,72 @@ pub fn parser_grammar() -> ParserGrammar {
             line_number: 35,
         },
         GrammarRule {
+            name: r#"method_with_block"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::Group { element: Box::new(GrammarElement::Alternation { choices: vec![
+                        GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
+                        GrammarElement::TokenReference { name: r#"KEYWORD"#.to_string() },
+                    ] }) },
+                GrammarElement::Optional { element: Box::new(GrammarElement::Sequence { elements: vec![
+                        GrammarElement::TokenReference { name: r#"LPAREN"#.to_string() },
+                        GrammarElement::Optional { element: Box::new(GrammarElement::Sequence { elements: vec![
+                                GrammarElement::RuleReference { name: r#"expression"#.to_string() },
+                                GrammarElement::Repetition { element: Box::new(GrammarElement::Sequence { elements: vec![
+                                        GrammarElement::TokenReference { name: r#"COMMA"#.to_string() },
+                                        GrammarElement::RuleReference { name: r#"expression"#.to_string() },
+                                    ] }) },
+                            ] }) },
+                        GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
+                    ] }) },
+                GrammarElement::RuleReference { name: r#"block"#.to_string() },
+            ] },
+            line_number: 52,
+        },
+        GrammarRule {
+            name: r#"block"#.to_string(),
+            body: GrammarElement::Alternation { choices: vec![
+                GrammarElement::RuleReference { name: r#"do_block"#.to_string() },
+                GrammarElement::RuleReference { name: r#"brace_block"#.to_string() },
+            ] },
+            line_number: 53,
+        },
+        GrammarRule {
+            name: r#"do_block"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::Literal { value: r#"do"#.to_string() },
+                GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"block_params"#.to_string() }) },
+                GrammarElement::Repetition { element: Box::new(GrammarElement::Sequence { elements: vec![
+                        GrammarElement::NegativeLookahead { element: Box::new(GrammarElement::Literal { value: r#"end"#.to_string() }) },
+                        GrammarElement::RuleReference { name: r#"statement"#.to_string() },
+                    ] }) },
+                GrammarElement::Literal { value: r#"end"#.to_string() },
+            ] },
+            line_number: 54,
+        },
+        GrammarRule {
+            name: r#"brace_block"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::TokenReference { name: r#"LBRACE"#.to_string() },
+                GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"block_params"#.to_string() }) },
+                GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"statement"#.to_string() }) },
+                GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
+            ] },
+            line_number: 55,
+        },
+        GrammarRule {
+            name: r#"block_params"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::Literal { value: r#"|"#.to_string() },
+                GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
+                GrammarElement::Repetition { element: Box::new(GrammarElement::Sequence { elements: vec![
+                        GrammarElement::TokenReference { name: r#"COMMA"#.to_string() },
+                        GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
+                    ] }) },
+                GrammarElement::Literal { value: r#"|"#.to_string() },
+            ] },
+            line_number: 56,
+        },
+        GrammarRule {
             name: r#"params"#.to_string(),
             body: GrammarElement::Sequence { elements: vec![
                 GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
@@ -84,7 +151,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
                     ] }) },
             ] },
-            line_number: 36,
+            line_number: 57,
         },
         GrammarRule {
             name: r#"if_statement"#.to_string(),
@@ -101,7 +168,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 37,
+            line_number: 58,
         },
         GrammarRule {
             name: r#"elsif_clause"#.to_string(),
@@ -115,7 +182,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 38,
+            line_number: 59,
         },
         GrammarRule {
             name: r#"else_clause"#.to_string(),
@@ -126,7 +193,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 39,
+            line_number: 60,
         },
         GrammarRule {
             name: r#"unless_statement"#.to_string(),
@@ -141,7 +208,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 40,
+            line_number: 61,
         },
         GrammarRule {
             name: r#"while_statement"#.to_string(),
@@ -154,7 +221,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 41,
+            line_number: 62,
         },
         GrammarRule {
             name: r#"until_statement"#.to_string(),
@@ -167,7 +234,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 42,
+            line_number: 63,
         },
         GrammarRule {
             name: r#"assignment"#.to_string(),
@@ -176,7 +243,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"EQUALS"#.to_string() },
                 GrammarElement::RuleReference { name: r#"expression"#.to_string() },
             ] },
-            line_number: 43,
+            line_number: 64,
         },
         GrammarRule {
             name: r#"method_call"#.to_string(),
@@ -195,12 +262,12 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
             ] },
-            line_number: 44,
+            line_number: 65,
         },
         GrammarRule {
             name: r#"expression_stmt"#.to_string(),
             body: GrammarElement::RuleReference { name: r#"expression"#.to_string() },
-            line_number: 45,
+            line_number: 66,
         },
         GrammarRule {
             name: r#"expression"#.to_string(),
@@ -214,7 +281,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"term"#.to_string() },
                     ] }) },
             ] },
-            line_number: 46,
+            line_number: 67,
         },
         GrammarRule {
             name: r#"term"#.to_string(),
@@ -228,7 +295,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"factor"#.to_string() },
                     ] }) },
             ] },
-            line_number: 47,
+            line_number: 68,
         },
         GrammarRule {
             name: r#"factor"#.to_string(),
@@ -246,7 +313,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
                 ] },
             ] },
-            line_number: 48,
+            line_number: 69,
         },
         GrammarRule {
             name: r#"symbol_literal"#.to_string(),
@@ -258,7 +325,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"STRING"#.to_string() },
                     ] }) },
             ] },
-            line_number: 49,
+            line_number: 70,
         },
         GrammarRule {
             name: r#"array_literal"#.to_string(),
@@ -273,7 +340,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACKET"#.to_string() },
             ] },
-            line_number: 50,
+            line_number: 71,
         },
         GrammarRule {
             name: r#"hash_literal"#.to_string(),
@@ -288,7 +355,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
             ] },
-            line_number: 51,
+            line_number: 72,
         },
         GrammarRule {
             name: r#"hash_entry"#.to_string(),
@@ -304,7 +371,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                 ] },
             ] },
-            line_number: 52,
+            line_number: 73,
         },
     ],
         version: 1,